refactor(core): rename some files name and methods
rename some files name and methods, fix some comments
